backend/requirements.txt
fastapi
uvicorn
websockets
google-genai>=1.0.0
python-dotenv
cloud-sql-python-connector[pymysql]
pymysql
sqlalchemy
google-auth
pyjwt
bcrypt
deepface
tf-keras
